+++ TanksDemo/Assets/Scripts/IDamage.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamage
 {
-    void TakeDamage(int _hitValue);
+    void TakeDamage(float _hitValue);
 }
